+++ private/buf/cmd/buf/command/registry/module/modulecommit/modulecommitinfo/usage.gen.go
@@ -14,6 +14,6 @@
 
 // Generated. DO NOT EDIT.
 
-package labelunarchive
+package modulecommitinfo
 
 import _ "github.com/bufbuild/buf/private/usage"