+++ private/buf/cmd/buf/command/registry/module/modulelabel/modulelabellist/usage.gen.go
@@ -14,6 +14,6 @@
 
 // Generated. DO NOT EDIT.
 
-package commitresolve
+package modulelabellist
 
 import _ "github.com/bufbuild/buf/private/usage"